generate_slope/requirements.txt
laspy>=2.0.0
numpy>=1.18.1
pyyaml>=6.0
tqdm>=4.45.0
